Copy Number and Single Base Substitution mutational signatures for 120 malignant pleural mesothelioma from the MESOMICS cohort

Copy Number and Single Base Substitution mutational signature attributions from Mangiante et al. Nat Genet 55, 607–618 (2023). https://doi.org/10.1038/s41588-023-01321-1

MESOMICS_CN_SBS

Format

MESOMICS_CN_SBS

A tibble with 120 rows and 41 columns:

ID_MESOMICS

Identifier of the patient

Sample

Identifier of the tumor sample

Sex

Patient sex

Age

Patient age at diagnosis

Smoking, Professional.Asbestos, ...

Patient exposure status

Survival.Censor, Survival.Time

Survival status (dead or alive) and time in months

Asbestos.Exposure.probability, Asbestos.Exposure.frequency, ...

Quantitative asbestos exposure variables

Type

Tumor histological type (WHO)

CNB

Copy number burden

CN1,...,CN19

Copy number signature attributions

SBS1,...,SBS40

Single Base Substitution signature attributions

Source

https://static-content.springer.com/esm/art%3A10.1038%2Fs41588-023-01321-1/MediaObjects/41588_2023_1321_MOESM4_ESM.xlsx