Compute the cosine similarity matrix for a set of mutational signatures.

Compute the cosine similarity matrix for a set of mutational signatures.

cossim(ref_sigs)

Arguments

ref_sigs

A matrix with K columns containing non-negative entries that sum to 1. Each column represents a signature (e.g., SBS1, SBS2), each row represents a mutation type (e.g., A[C>A]A), and each entry represents the abundance of that type in the signature.

Value

A K x K similarity matrix with diagonal elements equal to 1 and off-diagonal elements between 0 and 1. Entry S[j,k] is the similarity between signature j and signature k, equaling 1 if the categories are to be treated as identical and equaling 0 if they are to be treated as totally dissimilar. The default value is S = diag(K).

Examples

# Compute the cosine similarity matrix for the lung cancer
# in never smoker datasets from Zhang et al. 2021

 cossim(ref_sigs = as.matrix(Sherlock_LCINS_SBS.refs[,1:14]))
#>               SBS1        SBS2       SBS3       SBS5       SBS7a       SBS7b
#> SBS1   1.000000000 0.020180462 0.05355559 0.19372519 0.047661087 0.069351313
#> SBS2   0.020180462 1.000000000 0.09763503 0.25070396 0.714254359 0.266681901
#> SBS3   0.053555594 0.097635030 1.00000000 0.79310461 0.211375979 0.348041343
#> SBS5   0.193725195 0.250703962 0.79310461 1.00000000 0.359467381 0.430037417
#> SBS7a  0.047661087 0.714254359 0.21137598 0.35946738 1.000000000 0.721222110
#> SBS7b  0.069351313 0.266681901 0.34804134 0.43003742 0.721222110 1.000000000
#> SBS7c  0.024036005 0.058570636 0.32436409 0.25473791 0.083499680 0.086455249
#> SBS7d  0.014807097 0.072658879 0.27544233 0.33437332 0.071594589 0.123456468
#> SBS8   0.040327159 0.045182060 0.68991556 0.44581453 0.052160252 0.090836692
#> SBS13  0.006641537 0.016547762 0.28729716 0.15689968 0.017167311 0.014946282
#> SBS17a 0.007067398 0.001788106 0.24609497 0.20472739 0.006859816 0.022746199
#> SBS17b 0.002131456 0.001444865 0.11018468 0.09845749 0.002242365 0.005123663
#> SBS18  0.122234800 0.064847936 0.44076905 0.35593069 0.098792736 0.113974225
#> SBS40  0.040161906 0.331223252 0.88218056 0.83133419 0.397468930 0.411785275
#>             SBS7c      SBS7d       SBS8       SBS13      SBS17a      SBS17b
#> SBS1   0.02403600 0.01480710 0.04032716 0.006641537 0.007067398 0.002131456
#> SBS2   0.05857064 0.07265888 0.04518206 0.016547762 0.001788106 0.001444865
#> SBS3   0.32436409 0.27544233 0.68991556 0.287297156 0.246094971 0.110184675
#> SBS5   0.25473791 0.33437332 0.44581453 0.156899683 0.204727386 0.098457494
#> SBS7a  0.08349968 0.07159459 0.05216025 0.017167311 0.006859816 0.002242365
#> SBS7b  0.08645525 0.12345647 0.09083669 0.014946282 0.022746199 0.005123663
#> SBS7c  1.00000000 0.19966974 0.38626631 0.042338205 0.178990191 0.041710119
#> SBS7d  0.19966974 1.00000000 0.15975923 0.015843007 0.126075595 0.017124065
#> SBS8   0.38626631 0.15975923 1.00000000 0.069271577 0.130119910 0.019226160
#> SBS13  0.04233820 0.01584301 0.06927158 1.000000000 0.006642778 0.002715899
#> SBS17a 0.17899019 0.12607560 0.13011991 0.006642778 1.000000000 0.055045477
#> SBS17b 0.04171012 0.01712406 0.01922616 0.002715899 0.055045477 1.000000000
#> SBS18  0.09947618 0.08161410 0.63581197 0.128413437 0.028294197 0.031679590
#> SBS40  0.37133525 0.26971018 0.66139228 0.298896962 0.194267356 0.187639139
#>             SBS18      SBS40
#> SBS1   0.12223480 0.04016191
#> SBS2   0.06484794 0.33122325
#> SBS3   0.44076905 0.88218056
#> SBS5   0.35593069 0.83133419
#> SBS7a  0.09879274 0.39746893
#> SBS7b  0.11397423 0.41178527
#> SBS7c  0.09947618 0.37133525
#> SBS7d  0.08161410 0.26971018
#> SBS8   0.63581197 0.66139228
#> SBS13  0.12841344 0.29889696
#> SBS17a 0.02829420 0.19426736
#> SBS17b 0.03167959 0.18763914
#> SBS18  1.00000000 0.57888835
#> SBS40  0.57888835 1.00000000