Compute the within-sample diversity and across-sample heterogeneity of mutational signature activity in one or multiple populations of samples.

Across-sample variability is quantified by computing the population-genetic statistic Fst across the rows of the signature activity table. This method is based on the R package Fst, which implements an Fst-based assessment of Variability across vectors of relative Abundances. Mean within-sample variability is quantified by computing the mean Gini-Simpson index across all samples. Specify group if your table of signature activities contains mutiple populations of samples which you wish to compare. Specifying S allows sigvar to account for the pairwise cosine similarity among signatures. Specifying w or time allows for uneven weighting of the samples.

sigvar(
  sig_activity,
  K = NULL,
  group = NULL,
  S = NULL,
  w = NULL,
  time = NULL,
  normalized = FALSE
)

Arguments

sig_activity

A matrix or data frame with rows containing non-negative entries that sum to 1. Each row represents a sample, each column represents a mutational signature, and each entry represents the abundance of that signature in the sample. If sig_activity contains any metadata, it must be on the left-hand side of the matrix, the right K entries of each row must sum to 1, and K must be specified. Otherwise, all entries of each row must sum to 1.

K

Optional; an integer specifying the number of mutational signatures included in sig_activity. Default is K=ncol(sig_activity).

group

Optional; a string (or vector of strings) specifying the name(s) of the column(s) that describes which group(s) each sample belongs to. Use if sig_activity is a single matrix containing multiple groups of samples you wish to compare.

S

Optional; a K x K similarity matrix with diagonal elements equal to 1 and off-diagonal elements between 0 and 1. Entry S[j,k] is the similarity between signature j and signature k, equaling 1 if the categories are to be treated as identical and equaling 0 if they are to be treated as totally dissimilar. The default value is S = diag(K).

w

Optional; a vector of length nrow(sig_activity) with non-negative entries that sum to 1. Entry w[i] represents the weight placed on row i in the computation of the mean abundance of each category across rows. The default value is w = rep(1/nrow(sig_activity), nrow(sig_activity)).

time

Optional; a string specifying the name of the column that describes the sampling time for each row. Include if you wish to weight the variability measures by the distance between adjacent samples in the matrix.

normalized

Optional; should normalized Fst be used as the measure of across-sample variability? Default is normalized = FALSE; use normalized = TRUE to compute normalized Fst. Fst can only be normalized if it is not weighted.

Value

A data frame with columns corresponding to the across-sample variability and the mean within-sample variability for the provided samples.

Examples

# Compute the across-sample and mean within-sample variability of  mutational signatures
# in ESCC samples grouped by country
# We provide a cosine similarity matrix in order to account for cosine similarity among signatures
 sigvar(sig_activity = smoker_sigs_chen, K = 3, group = "Smoker", S = smoker_sigs_chen_cossim)
#> Joining with `by = join_by(Smoker)`
#>       Smoker across_sample_heterogeneity mean_within_sample_diversity
#> 1     Smoker                   0.2308918                    0.3425499
#> 2 Non-smoker                   0.3055201                    0.2810333